@@ -7,36 +7,39 @@
     "ng": "ng",
     "start": "ng serve",
     "build": "ng build",
-    "test": "ng test",
+    "test": "jest",
     "lint": "ng lint",
     "e2e": "ng e2e"
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "~7.2.0",
-    "@angular/core": "~7.2.0",
-    "@angular/fire": "^5.1.1",
-    "@angular/forms": "~7.2.0",
-    "@angular/http": "~7.2.0",
-    "@angular/platform-browser": "~7.2.0",
-    "@angular/platform-browser-dynamic": "~7.2.0",
-    "@angular/pwa": "^0.11.4",
-    "@angular/router": "~7.2.0",
-    "@angular/service-worker": "~7.2.0",
-    "@ionic-native/camera": "5.0.0",
-    "@ionic-native/core": "5.0.0",
-    "@ionic-native/file": "5.0.0",
-    "@ionic-native/firebase": "5.0.0",
-    "@ionic-native/image-resizer": "^5.0.0",
-    "@ionic-native/ionic-webview": "^5.0.0",
-    "@ionic-native/network": "^5.0.0",
-    "@ionic-native/splash-screen": "5.0.0",
-    "@ionic-native/status-bar": "5.0.0",
-    "@ionic/angular": "4.0.0-rc.2",
-    "@ionic/storage": "^2.2.0",
+    "@angular-devkit/schematics-cli": "^0.800.3",
+    "@angular/common": "~8.2.2",
+    "@angular/core": "~8.2.2",
+    "@angular/fire": "5.2.1",
+    "@angular/forms": "~8.2.2",
+    "@angular/http": "7.2.15",
+    "@angular/platform-browser": "~8.2.2",
+    "@angular/platform-browser-dynamic": "~8.2.2",
+    "@angular/pwa": "^0.800.3",
+    "@angular/router": "~8.2.2",
+    "@angular/service-worker": "~8.2.2",
+    "@ionic-native/camera": "5.8.0",
+    "@ionic-native/core": "5.8.0",
+    "@ionic-native/file": "5.8.0",
+    "@ionic-native/firebase": "5.8.0",
+    "@ionic-native/image-resizer": "5.8.0",
+    "@ionic-native/ionic-webview": "5.8.0",
+    "@ionic-native/network": "5.8.0",
+    "@ionic-native/splash-screen": "5.8.0",
+    "@ionic-native/status-bar": "5.8.0",
+    "@ionic/angular": "4.7.4",
+    "@ionic/storage": "2.2.0",
     "@types/chai": "^4.1.7",
     "@types/cucumber": "^4.0.4",
+    "@wdio/cli": "^5.4.13",
     "chai": "^4.2.0",
+    "cordova": "^9.0.0",
     "cordova-browser": "5.0.4",
     "cordova-ios": "4.5.5",
     "cordova-plugin-camera": "^4.0.3",
@@ -52,40 +55,58 @@
     "cordova-sqlite-storage": "^2.6.0",
     "core-js": "^2.6.2",
     "firebase": "^5.8.0",
-    "firebase-tools": "^6.3.1",
+    "firebase-tools": "7.2.2",
     "guid-typescript": "^1.0.9",
     "info.protonet.imageresizer": "0.1.1",
     "protractor-cucumber-framework": "^6.1.1",
-    "rxjs": "~6.3.3",
-    "zone.js": "^0.8.28"
+    "rxjs": "~6.5.2",
+    "zone.js": "^0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "~0.11.4",
-    "@angular-devkit/build-angular": "^0.12.2",
-    "@angular-devkit/core": "~7.2.0",
-    "@angular-devkit/schematics": "~7.2.0",
-    "@angular/cli": "~7.2.0",
-    "@angular/compiler": "~7.2.0",
-    "@angular/compiler-cli": "~7.2.0",
-    "@angular/language-service": "~7.2.0",
-    "@ionic/angular-toolkit": "~1.2.0",
+    "@angular-builders/jest": "^8.0.2",
+    "@angular-devkit/architect": "^0.802.2",
+    "@angular-devkit/build-angular": "^0.802.2",
+    "@angular-devkit/core": "^8.2.2",
+    "@angular-devkit/schematics": "^8.2.2",
+    "@angular/cli": "~8.2.2",
+    "@angular/compiler": "~8.2.2",
+    "@angular/compiler-cli": "~8.2.2",
+    "@angular/language-service": "~8.2.2",
+    "@babel/plugin-syntax-dynamic-import": "^7.2.0",
+    "@babel/preset-env": "^7.4.5",
+    "@ionic/angular-toolkit": "1.5.1",
     "@ionic/lab": "1.0.16",
-    "@types/jasmine": "^2.8.15",
-    "@types/jasminewd2": "~2.0.3",
+    "@types/jest": "^24.0.15",
     "@types/node": "~10.12.0",
+    "@wdio/local-runner": "^5.4.14",
     "codelyzer": "~4.5.0",
     "cucumber": "^5.1.0",
-    "jasmine-core": "~2.99.1",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~3.1.4",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.4.0",
+    "jest": "^24.8.0",
+    "jest-preset-angular": "^7.1.1",
     "ts-node": "~7.0.0",
     "tslint": "^5.12.1",
-    "typescript": "~3.1.6"
+    "typescript": "~3.5.3",
+    "wdio-webcomponents": "^1.1.0"
+  },
+  "jest": {
+    "preset": "jest-preset-angular",
+    "setupFilesAfterEnv": [
+      "<rootDir>/setupJest.ts"
+    ],
+    "testPathIgnorePatterns": [
+      "<rootDir>/node_modules",
+      "<rootDir>/dist/",
+      "<rootDir>/src/test.ts"
+    ],
+    "globals": {
+      "ts-jest": {
+        "tsConfig": "<rootDir>/tsconfig.spec.json",
+        "stringifyContentPathRegex": "\\.html$",
+        "asTransformers": [
+          "<rootDir>/node_modules/jest-preset-angular/InlineHtmlStripStylesTransformer"
+        ]
+      }
+    }
   },
   "description": "Simple shopping list and recipe helper!",
   "cordova": {
