@@ -1,23 +1,29 @@
-import { browser, by, element } from 'protractor';
+import {browser, by, element, protractor} from 'protractor';
 
 export class SignInPage {
+
 
     navigateTo() {
         return browser.get('/sign-in');
     }
 
     getText() {
-        return element(by.id('no-shopping-list-text')).getText();
+        return element(by.id('shoppingListTitle')).getText();
     }
 
     enterUsername(username: string) {
-        element(by.id('username')).sendKeys(username);
+        element(by.css('ion-input[id=username] > input')).sendKeys(username);
     }
     enterPassword(password: string) {
-        element(by.id('password')).sendKeys(password);
+        element(by.css('ion-input[name=password] > input')).sendKeys(password);
     }
     pushLogin() {
-        element(by.id('login-button')).click();
-        browser.pause();
+        element(by.css('ion-button[id=login-button]')).click();
+    }
+
+    pushLogout() {
+        if (element(by.css('ion-button[id=logout-button]')).isPresent()) {
+            element(by.css('ion-button[id=logout-button]')).click();
+        }
     }
 }
